extract animated counter component to their own file
